Include the activation_id in the Release signal
This fixes a potential race condition where an input capture gets
Deactivated and immediately Activates again while the client attempts to
Release the input capture. By including the activation_id the compositor
can ignore the Release and instead wait on the client to send a new
Release with the now-updated activation_id.